+++ web/src/routes/(user)/workflows/[workflowId]/+page.svelte
@@ -13,7 +13,6 @@
     ActionBar,
     AppShell,
     AppShellBar,
-    Badge,
     Button,
     Card,
     CardBody,
@@ -36,23 +35,20 @@
   } from '@immich/ui';
   import {
     mdiArrowLeft,
-    mdiAutoFix,
     mdiCodeJson,
     mdiContentSave,
-    mdiDragVertical,
-    mdiFilterVariant,
     mdiFlashOutline,
     mdiFormatListBulletedSquare,
     mdiInformationOutline,
     mdiPencilOutline,
     mdiPlus,
-    mdiTrashCanOutline,
   } from '@mdi/js';
   import { cloneDeep, isEqual } from 'lodash-es';
   import { flushSync } from 'svelte';
   import { t } from 'svelte-i18n';
   import type { PageData } from './$types';
   import WorkflowJsonEditor from './WorkflowJsonEditor.svelte';
+  import WorkflowStepCard from './WorkflowStepCard.svelte';
   import WorkflowStepDragImage from './WorkflowStepDragImage.svelte';
   import WorkflowSummary from './WorkflowSummary.svelte';
 
@@ -88,12 +84,6 @@
 
   const workflowSummary = $derived({ name, description, trigger, steps });
   const workflowJsonContent = $derived<WorkflowJsonContent>({ description, enabled, name, steps, trigger });
-  const stepsWithConfigEntries = $derived(
-    steps.map((step) => ({
-      step,
-      configEntries: getConfigEntries(step.config),
-    })),
-  );
 
   const hasChanges = $derived(
     enabled !== savedWorkflow.enabled ||
@@ -126,7 +116,7 @@
     }
   };
 
-  const handleDragStart = (index: number) => (event: DragEvent) => {
+  const handleDragStart = (index: number, event: DragEvent) => {
     draggedIndex = index;
     if (event.dataTransfer) {
       event.dataTransfer.effectAllowed = 'move';
@@ -148,7 +138,7 @@
     }
   };
 
-  const handleDragOver = (index: number) => (event: DragEvent) => {
+  const handleDragOver = (index: number, event: DragEvent) => {
     if (draggedIndex === null) {
       return;
     }
@@ -161,13 +151,13 @@
     }
   };
 
-  const handleDragLeave = (index: number) => () => {
+  const handleDragLeave = (index: number) => {
     if (dropTargetIndex === index) {
       dropTargetIndex = null;
     }
   };
 
-  const handleDrop = (index: number) => (event: DragEvent) => {
+  const handleDrop = (index: number, event: DragEvent) => {
     event.preventDefault();
     const from = draggedIndex;
     draggedIndex = null;
@@ -203,39 +193,6 @@
   };
 
   const onClose = () => goto(Route.workflows());
-
-  const truncate = (input: string, max = 24) => (input.length > max ? input.slice(0, max - 1) + '…' : input);
-
-  const formatConfigValue = (value: unknown): string => {
-    if (value === null || value === undefined) {
-      return '—';
-    }
-    if (typeof value === 'boolean') {
-      return value ? 'on' : 'off';
-    }
-    if (typeof value === 'number') {
-      return String(value);
-    }
-    if (typeof value === 'string') {
-      return `"${truncate(value)}"`;
-    }
-    if (Array.isArray(value)) {
-      if (value.length === 0) {
-        return $t('none');
-      }
-      const items = value.map((v) => (v !== null && typeof v === 'object' ? '{…}' : String(v)));
-      const joined = items.join(' · ');
-      if (joined.length <= 28) {
-        return `"${joined}"`;
-      }
-      return $t('items_count', { values: { count: value.length } });
-    }
-    return '{…}';
-  };
-
-  function getConfigEntries(config: WorkflowStepDto['config']) {
-    return Object.entries(config ?? {}).filter(([, value]) => value !== null && value !== undefined && value !== '');
-  }
 
   const onChangeTrigger = async () => {
     const newTrigger = await modalManager.show(WorkflowTriggerPicker, { selected: trigger });
@@ -435,105 +392,24 @@
           </div>
         {/snippet}
 
-        {#each stepsWithConfigEntries as { step, configEntries }, index (index)}
-          {@const method = pluginManager.getMethod(step.method)}
-          {@const isFilter = method?.uiHints?.includes('filter') ?? false}
-          {@const isDragging = draggedIndex === index}
-          {@const isDragHandleHovered = dragHandleHoverIndex === index}
-          {@const isDropTarget = dropTargetIndex === index && draggedIndex !== null && draggedIndex !== index}
+        {#each steps as step, index (index)}
           {@render sequenceConnector()}
-          <div
-            class="w-full transition-all"
-            class:opacity-40={isDragging}
-            class:scale-[0.99]={isDragging}
-            ondragover={handleDragOver(index)}
-            ondragleave={handleDragLeave(index)}
-            ondrop={handleDrop(index)}
-            role="listitem"
-          >
-            <Card
-              class="shadow-none transition-colors {isDropTarget
-                ? 'border-primary ring-2 ring-primary-200'
-                : isDragHandleHovered
-                  ? 'border-dashed border-primary'
-                  : ''}"
-            >
-              <CardHeader>
-                <div class="flex items-center gap-2">
-                  <!-- svelte-ignore a11y_no_static_element_interactions -->
-                  <div
-                    class="flex shrink-0 cursor-grab items-center justify-center rounded-md border border-transparent p-1 text-light-400 select-none hover:border-primary-200 hover:bg-primary-50 hover:text-primary active:cursor-grabbing"
-                    aria-label={$t('drag_to_reorder')}
-                    draggable="true"
-                    onmouseenter={() => (dragHandleHoverIndex = index)}
-                    onmouseleave={() => (dragHandleHoverIndex = null)}
-                    ondragstart={handleDragStart(index)}
-                    ondragend={handleDragEnd}
-                    title={$t('drag_to_reorder')}
-                  >
-                    <Icon icon={mdiDragVertical} size="20" />
-                  </div>
-                  <div
-                    class="flex size-10 shrink-0 items-center justify-center rounded-lg"
-                    class:bg-primary-50={isFilter}
-                    class:bg-warning-50={!isFilter}
-                  >
-                    <Icon
-                      icon={isFilter ? mdiFilterVariant : mdiAutoFix}
-                      size="20"
-                      class={isFilter ? 'text-primary' : 'text-warning'}
-                    />
-                  </div>
-                  <div class="flex min-w-0 flex-1 flex-col">
-                    <CardTitle class="truncate">
-                      <span class="mr-1 font-bold text-light-500">{index + 1}</span>
-                      {pluginManager.getMethodLabel(step.method)}
-                    </CardTitle>
-                    {#if method?.description}
-                      <CardDescription class="truncate">{method.description}</CardDescription>
-                    {/if}
-                  </div>
-                  <div class="flex shrink-0 items-center gap-1">
-                    <IconButton
-                      icon={mdiPencilOutline}
-                      aria-label={$t('edit')}
-                      variant="ghost"
-                      shape="round"
-                      color="secondary"
-                      size="small"
-                      onclick={() => handleEditStep(index)}
-                    />
-                    <IconButton
-                      icon={mdiTrashCanOutline}
-                      aria-label={$t('delete')}
-                      variant="ghost"
-                      shape="round"
-                      color="danger"
-                      size="small"
-                      onclick={() => handleDeleteStep(index)}
-                    />
-                  </div>
-                </div>
-              </CardHeader>
-
-              {#if configEntries.length > 0}
-                <CardBody class="py-3">
-                  <div class="flex flex-wrap items-center gap-1.5">
-                    {#each configEntries as [key, value] (key)}
-                      <Badge
-                        color={isFilter ? 'info' : 'warning'}
-                        shape="round"
-                        size="small"
-                        class="border font-mono {isFilter ? 'border-primary-200' : 'border-warning-200'}"
-                      >
-                        <span class="opacity-60">{key}</span>{formatConfigValue(value)}
-                      </Badge>
-                    {/each}
-                  </div>
-                </CardBody>
-              {/if}
-            </Card>
-          </div>
+          <WorkflowStepCard
+            {step}
+            {index}
+            isDragging={draggedIndex === index}
+            isDragHandleHovered={dragHandleHoverIndex === index}
+            isDropTarget={dropTargetIndex === index && draggedIndex !== null && draggedIndex !== index}
+            onEdit={handleEditStep}
+            onDelete={handleDeleteStep}
+            onDragStart={handleDragStart}
+            onDragEnd={handleDragEnd}
+            onDragOver={handleDragOver}
+            onDragLeave={handleDragLeave}
+            onDrop={handleDrop}
+            onDragHandleEnter={(i) => (dragHandleHoverIndex = i)}
+            onDragHandleLeave={() => (dragHandleHoverIndex = null)}
+          />
         {/each}
 
         <Button
